requirements.txt
# =============================================================================
# Antigravity ML Dependencies
# =============================================================================

# Core ML Frameworks
torch>=2.1.0
transformers>=4.35.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization & Observation
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.15.0

# Model Validation & Serialization
pyyaml>=6.0.1
onnx>=1.15.0

# File Watching (Python alternative for watchers)
watchdog>=3.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.1

# Hugging Face Ecosystem
datasets>=2.14.0
accelerate>=0.24.0
